[mtouch/mmp] Split out the RunCommand[Async] methods to a separate file so that the generator can reuse more easily.
